[AE] Drop CAEFactory
