+++ xbmc/cores/AudioEngine/Sinks/AESinkPULSE.h
@@ -22,7 +22,6 @@
 #include "system.h"
 
 #include "cores/AudioEngine/Interfaces/AESink.h"
-#include "cores/AudioEngine/AEFactory.h"
 #include "Utils/AEDeviceInfo.h"
 #include "Utils/AEUtil.h"
 #include <pulse/pulseaudio.h>
