+++ xbmc/cores/AudioEngine/Sinks/alsa/ALSAHControlMonitor.cpp
@@ -21,11 +21,11 @@
 #include "system.h"
 #ifdef HAS_ALSA
 
+#include "cores/AudioEngine/Engines/ActiveAE/ActiveAE.h"
 #include "ALSAHControlMonitor.h"
-
-#include "AEFactory.h"
 #include "linux/FDEventMonitor.h"
 #include "utils/log.h"
+#include "ServiceBroker.h"
 
 CALSAHControlMonitor::CALSAHControlMonitor()
 {
@@ -131,7 +131,7 @@ int CALSAHControlMonitor::HCTLCallback(snd_hctl_elem_t *elem, unsigned int mask)
      * Currently we just re-enumerate on any change.
      * Custom callbacks for handling other control monitoring may be implemented when needed.
      */
-    CAEFactory::DeviceChange();
+    CServiceBroker::GetActiveAE().DeviceChange();
   }
 
   return 0;
